feat: limit size of string to int ICIJ/datashare#894

+++ src/main/java/org/icij/datashare/text/indexing/ExtractedText.java
@@ -2,12 +2,12 @@ package org.icij.datashare.text.indexing;
 
 public class ExtractedText {
     public final String content;
-    public final long maxOffset;
-    public final long offset;
-    public final long limit;
+    public final int maxOffset;
+    public final int offset;
+    public final int limit;
 
 
-    public ExtractedText(String content, long offset, long limit, long maxOffset) {
+    public ExtractedText(String content, int offset, int limit, int maxOffset) {
         this.content = content;
         this.offset = offset;
         this.limit = limit;
